Added configure check for kind parameter in shape intrinsic
We need this feature to safely read large datasets in HDF5 snapshots.
When using a compiler which doesn't support it we will still be able
to read snapshots where dataset sizes are < 2^31.

TODO: add a check for the case where we don't have the kind parameter
but we are trying to read a large dataset.